Move types, add fetch finalUrl return
